Updated and reimÃplemented a lot of util classes and implemented aisc myProfile logic
I refactored a lot of methods in the util package and redid the registration process logic. The resources manager now
caches already loaded resources locally and same with the account manager (which is responsible for loading other users
accounts and managing the current user account). I also implemented update requests for the current user account. PrivateProfile was
renamed to ContactDetails and got an own fragment, in case the locally stored ContactDetails get lost.